Refactor label mode store and modularize components
